# videodna-studio/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VideoDNA Studio API",
    "description": "API do VideoDNA Studio — análise, desmontagem semântica, edição estrutural e reconstrução de vídeos com múltiplos modelos de IA. Erros seguem o formato `{error: {code, message, details, requestId}}` com códigos normalizados.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Listar projetos",
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "type": "array",
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Criar projeto",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/demo": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Criar um projeto de exemplo (vídeo de demonstração) e iniciar a análise",
        "operationId": "create_demo_project_projects_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DemoProjectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Detalhes do projeto",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Atualizar nome/locks/modo",
        "operationId": "update_project_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Excluir projeto e toda a mídia",
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/media": {
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Excluir apenas a mídia do projeto",
        "operationId": "delete_media_projects__project_id__media_delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/duplicate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Duplicar projeto",
        "operationId": "duplicate_project_projects__project_id__duplicate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "pipelineVersion": {
            "type": "string",
            "title": "Pipelineversion"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "lowConfidenceCount": {
            "type": "integer",
            "title": "Lowconfidencecount",
            "default": 0
          },
          "reusedFromId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reusedfromid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "pipelineVersion",
          "mock",
          "summary",
          "lowConfidenceCount",
          "reusedFromId",
          "createdAt",
          "completedAt"
        ],
        "title": "AnalysisOut"
      },
...
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "AnalysisStatus"
      },
...
      "DemoProjectOut": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectOut"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reused": {
            "type": "boolean",
            "title": "Reused",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "project",
          "job",
          "reused"
        ],
        "title": "DemoProjectOut"
      },
      "DemoProjectRequest": {
        "properties": {
          "fresh": {
            "type": "boolean",
            "title": "Fresh",
            "default": false
          }
        },
        "type": "object",
        "title": "DemoProjectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "INGEST",
          "ANALYSIS",
          "PREVIEW",
          "GENERATION"
        ],
        "title": "JobKind"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "title": "Projectid"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "planId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planid"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "kind",
          "status",
          "stage",
          "progress",
          "message",
          "errorCode",
          "errorMessage",
          "result",
          "planId",
          "attempts",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "title": "JobOut"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "PREPARING",
          "ANALYZING",
          "GENERATING",
          "QA",
          "REPAIRING",
          "ASSEMBLING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobStatus"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSettings-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
      "ProjectLocks-Input": {
        "properties": {
          "story": {
            "type": "boolean",
            "title": "Story",
            "default": true
          },
          "camera": {
            "type": "boolean",
            "title": "Camera",
            "default": true
          },
          "motion": {
            "type": "boolean",
            "title": "Motion",
            "default": true
          },
          "audio": {
            "type": "boolean",
            "title": "Audio",
            "default": true
          },
          "timing": {
            "type": "boolean",
            "title": "Timing",
            "default": true
          }
        },
        "type": "object",
        "title": "ProjectLocks"
      },
      "ProjectLocks-Output": {
        "properties": {
          "story": {
            "type": "boolean",
            "title": "Story",
            "default": true
          },
          "camera": {
            "type": "boolean",
            "title": "Camera",
            "default": true
          },
          "motion": {
            "type": "boolean",
            "title": "Motion",
            "default": true
          },
          "audio": {
            "type": "boolean",
            "title": "Audio",
            "default": true
          },
          "timing": {
            "type": "boolean",
            "title": "Timing",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "story",
          "camera",
          "motion",
          "audio",
          "timing"
        ],
        "title": "ProjectLocks"
      },
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettings-Output"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "sourceVideo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceVideoOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "activeJob": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "latestOutputId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestoutputid"
          },
          "editCount": {
            "type": "integer",
            "title": "Editcount",
            "default": 0
          },
          "totalCost": {
            "type": "number",
            "title": "Totalcost",
            "default": 0.0
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "BRL"
          },
          "isDemo": {
            "type": "boolean",
            "title": "Isdemo",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "settings",
          "createdAt",
          "updatedAt",
          "sourceVideo",
          "analysis",
          "activeJob",
          "latestOutputId",
          "editCount",
          "totalCost",
          "currency",
          "isDemo"
        ],
        "title": "ProjectOut"
      },
      "ProjectSettings-Input": {
        "properties": {
          "locks": {
            "$ref": "#/components/schemas/ProjectLocks-Input",
            "default": {
              "story": true,
              "camera": true,
              "motion": true,
              "audio": true,
              "timing": true
            }
          },
          "qualityMode": {
            "$ref": "#/components/schemas/QualityMode",
            "default": "BALANCED"
          },
          "editorMode": {
            "type": "string",
            "enum": [
              "MANUAL",
              "SMART",
              "AUTO"
            ],
            "title": "Editormode",
            "default": "MANUAL"
          }
        },
        "type": "object",
        "title": "ProjectSettings"
      },
      "ProjectSettings-Output": {
        "properties": {
          "locks": {
            "$ref": "#/components/schemas/ProjectLocks-Output",
            "default": {
              "story": true,
              "camera": true,
              "motion": true,
              "audio": true,
              "timing": true
            }
          },
          "qualityMode": {
            "$ref": "#/components/schemas/QualityMode",
            "default": "BALANCED"
          },
          "editorMode": {
            "type": "string",
            "enum": [
              "MANUAL",
              "SMART",
              "AUTO"
            ],
            "title": "Editormode",
            "default": "MANUAL"
          }
        },
        "type": "object",
        "required": [
          "locks",
          "qualityMode",
          "editorMode"
        ],
        "title": "ProjectSettings"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "UPLOADING",
          "UPLOADED",
          "ANALYZING",
          "READY",
          "GENERATING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "ProjectStatus"
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSettings-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "QualityMode": {
        "type": "string",
        "enum": [
          "ECONOMY",
          "BALANCED",
          "MAX"
        ],
        "title": "QualityMode"
      },
...
      "SourceVideoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "originalFilename": {
            "type": "string",
            "title": "Originalfilename"
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype"
          },
          "sizeBytes": {
            "type": "integer",
            "title": "Sizebytes"
          },
          "status": {
            "$ref": "#/components/schemas/SourceVideoStatus"
          },
          "durationSec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationsec"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "rightsConfirmedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rightsconfirmedat"
          },
          "proxyUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxyurl"
          },
          "posterUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posterurl"
          },
          "retentionUntil": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retentionuntil"
          }
        },
        "type": "object",
        "required": [
          "id",
          "originalFilename",
          "contentType",
          "sizeBytes",
          "status",
          "durationSec",
          "width",
          "height",
          "fps",
          "errorCode",
          "rightsConfirmedAt",
          "proxyUrl",
          "posterUrl",
          "retentionUntil"
        ],
        "title": "SourceVideoOut"
      },
      "SourceVideoStatus": {
        "type": "string",
        "enum": [
          "PENDING_UPLOAD",
          "UPLOADED",
          "INGESTING",
          "READY",
          "REJECTED",
          "DELETED"
        ],
        "title": "SourceVideoStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}